Done Gettext MO file parser
git-svn-id: https://appetizer.googlecode.com/svn@394 ff61cf6c-2836-11df-8847-f3d5e5ef3dde

+++ branches/QtVersion/Application.cpp
@@ -103,7 +103,7 @@ void Application::loadFolderItems() {
 
   if (fileVersion == "1.0") {
     FolderItemVector groups = rootFolderItem_->detachAllGroups();
-    for (int i = 0; i < groups.size(); i++) {
+    for (int i = 0; i < (int)groups.size(); i++) {
       FolderItem* group = groups.at(i);
       group->convertGroupToSection();
       rootFolderItem_->addChild(group);
